+++ src/integrators/photonmapper/bre.h
@@ -55,7 +55,7 @@ protected:
 	virtual ~BeamRadianceEstimator();
 
 	/// Fit a hierarchy of bounding boxes to the stored photons
-	AABB buildHierarchy(size_t index);
+	AABB buildHierarchy(uint32_t index);
 
 	/// Heap convenience routines
 	inline uint32_t leftChild(uint32_t index) const { return 2*index; }
@@ -70,9 +70,9 @@ private:
 	};
 
 	BRENode *m_nodes;
-	size_t m_photonCount;
-	size_t m_lastInnerNode;
-	size_t m_lastRChildNode;
+	uint32_t m_photonCount;
+	uint32_t m_lastInnerNode;
+	uint32_t m_lastRChildNode;
 	int m_depth;
 	Float m_scaleFactor;
 };
